Split constants out based on platform.
