+++ lib/sys/unix/sys/filesystem.rb
@@ -13,32 +13,6 @@ module Sys
 
     private_class_method :new
 
-    # Readable versions of constant names
-    OPT_NAMES = {
-      MNT_RDONLY => 'read-only',
-      MNT_SYNCHRONOUS => 'synchronous',
-      MNT_NOEXEC => 'noexec',
-      MNT_NOSUID => 'nosuid',
-      MNT_NODEV => 'nodev',
-      MNT_UNION => 'union',
-      MNT_ASYNC => 'asynchronous',
-      MNT_CPROTECT => 'content-protection',
-      MNT_EXPORTED => 'exported',
-      MNT_QUARANTINE => 'quarantined',
-      MNT_LOCAL => 'local',
-      MNT_QUOTA => 'quotas',
-      MNT_ROOTFS => 'rootfs',
-      MNT_DONTBROWSE => 'nobrowse',
-      MNT_IGNORE_OWNERSHIP => 'noowners',
-      MNT_AUTOMOUNTED => 'automounted',
-      MNT_JOURNALED => 'journaled',
-      MNT_NOUSERXATTR => 'nouserxattr',
-      MNT_DEFWRITE => 'defwrite',
-      MNT_NOATIME => 'noatime'
-    }.freeze
-
-    private_constant :OPT_NAMES
-
     # File used to read mount informtion from.
     if File.exist?('/etc/mtab')
       MOUNT_FILE = '/etc/mtab'.freeze
@@ -318,7 +292,7 @@ module Sys
           string = ''
           flags = mnt[:f_flags] & MNT_VISFLAGMASK
 
-          OPT_NAMES.each do |key, val|
+          Constants::MOUNT_OPTION_NAMES.each do |key, val|
             if flags & key > 0
               if string.empty?
                 string += val
