+++ lib/sys/unix/sys/filesystem/constants/dragonfly.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require_relative 'generic'
